-removed debug and release folders from github tracking -moved glad to its own folder -x86 builds currently not working, maybe due to glfw or imgui -x64 builds do work

+++ src/framework/rendering/shaderClass.cpp
@@ -12,7 +12,7 @@ std::string GetShaderFromFile(const char* filename)
 	{
 		std::string contents;
 		in.seekg(0, std::ios::end);
-		contents.resize(in.tellg());
+		contents.resize((const unsigned int)in.tellg());
 		in.seekg(0, std::ios::beg);
 		in.read(&contents[0], contents.size());
 		in.close();
